Refactoring next step, keep going;

+++ app/src/main/java/tk/lorddarthart/weathertest/util/converter/SimpleWeatherPlusMinusConverter.kt
@@ -1,7 +1,7 @@
 package tk.lorddarthart.weathertest.util.converter
 
 object SimpleWeatherPlusMinusConverter {
-    fun getPlusInFront(temp: Double): String {
+    fun getPlusInFront(temp: Int): String {
         return if (temp > 0.0) {
             "+$temp"
         } else {
